k230dev-yolo/workspace/requirements-convert.txt
# Conversion Environment Requirements
# For use with the 'convert' Docker service

# ONNX Processing
onnx==1.15.0
onnxruntime==1.16.3

# nncase for K230D
# Note: Versions may need adjustment based on K230D SDK
nncase==2.10.0
nncase-kpu==2.10.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0

# Data Processing
numpy==1.24.3

# Configuration
pyyaml==6.0.1
tqdm==4.66.1

# Utilities
scipy==1.11.4
